refactor(api): move DifferentExternalIdentifier error
to identity-access-management bounded context
