+++ api/src/identity-access-management/domain/usecases/find-user-for-oidc-reconciliation.usecase.js
@@ -1,5 +1,4 @@
-import { DifferentExternalIdentifierError } from '../../../../lib/domain/errors.js';
-import { AuthenticationKeyExpired } from '../errors.js';
+import { AuthenticationKeyExpired, DifferentExternalIdentifierError } from '../errors.js';
 
 /**
  * @typedef {function} findUserForOidcReconciliation
