+++ api/tests/identity-access-management/unit/domain/usecases/find-user-for-oidc-reconciliation.usecase.test.js
@@ -1,5 +1,7 @@
-import { DifferentExternalIdentifierError } from '../../../../../lib/domain/errors.js';
-import { AuthenticationKeyExpired } from '../../../../../src/identity-access-management/domain/errors.js';
+import {
+  AuthenticationKeyExpired,
+  DifferentExternalIdentifierError,
+} from '../../../../../src/identity-access-management/domain/errors.js';
 import { findUserForOidcReconciliation } from '../../../../../src/identity-access-management/domain/usecases/find-user-for-oidc-reconciliation.usecase.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
