+++ api/tests/unit/domain/usecases/reconcile-oidc-user-for-admin_test.js
@@ -1,7 +1,9 @@
-import { DifferentExternalIdentifierError } from '../../../../lib/domain/errors.js';
 import { AuthenticationMethod } from '../../../../lib/domain/models/index.js';
 import { reconcileOidcUserForAdmin } from '../../../../lib/domain/usecases/reconcile-oidc-user-for-admin.js';
-import { AuthenticationKeyExpired } from '../../../../src/identity-access-management/domain/errors.js';
+import {
+  AuthenticationKeyExpired,
+  DifferentExternalIdentifierError,
+} from '../../../../src/identity-access-management/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | reconcile-oidc-user-for-admin', function () {
